Refactor to extract code duplication
Fixes these bugs, too:
- Polygon's simplify method didn't square the tolerance parameter when it should have
- Polygon's 'highestQuality = false' didn't work at all(!)
