+++ Sources/Turf/Geometries/LineString.swift
@@ -216,96 +216,6 @@ extension LineString {
         return closestCoordinate
     }
 
-    private func squareDistance(from origin: LocationCoordinate2D, to destination: LocationCoordinate2D) -> Double {
-        let dx = origin.longitude - destination.longitude
-        let dy = origin.latitude - destination.latitude
-        return dx * dx + dy * dy
-    }
-
-    private mutating func simplified(radialTolerance: Double) {
-        guard coordinates.count > 2 else { return }
-
-        var prevCoordinate = coordinates[0]
-        var newCoordinates = [prevCoordinate]
-        var coordinate = coordinates[1]
-
-        for index in 1 ..< coordinates.count {
-            coordinate = coordinates[index]
-
-            if squareDistance(from: coordinate, to: prevCoordinate) > radialTolerance {
-                newCoordinates.append(coordinate)
-                prevCoordinate = coordinate
-            }
-        }
-
-        if prevCoordinate != coordinate {
-            newCoordinates.append(coordinate)
-        }
-
-        coordinates = newCoordinates
-    }
-
-    private func squareSegmentDistance(_ coordinate: LocationCoordinate2D, segmentStart: LocationCoordinate2D, segmentEnd: LocationCoordinate2D) -> LocationDistance {
-
-        var x = segmentStart.latitude
-        var y = segmentStart.longitude
-        var dx = segmentEnd.latitude - x
-        var dy = segmentEnd.longitude - y
-
-        if dx != 0 || dy != 0 {
-            let t = ((coordinate.latitude - x) * dx + (coordinate.longitude - y) * dy) / (dx * dx + dy * dy)
-            if t > 1 {
-                x = segmentEnd.latitude
-                y = segmentEnd.longitude
-            } else if t > 0 {
-                x += dx * t
-                y += dy * t
-            }
-        }
-
-        dx = coordinate.latitude - x
-        dy = coordinate.longitude - y
-
-        return dx * dx + dy * dy
-    }
-
-    private func simplifyDouglasPeuckerStep(_ coordinates: [LocationCoordinate2D], first: Int, last: Int, tolerance: Double, simplified: inout [LocationCoordinate2D]) {
-
-        var maxSquareDistance = tolerance
-        var index = 0
-
-        for i in first + 1 ..< last {
-            let squareDistance = squareSegmentDistance(coordinates[i], segmentStart: coordinates[first], segmentEnd: coordinates[last])
-
-            if squareDistance > maxSquareDistance {
-                index = i
-                maxSquareDistance = squareDistance
-            }
-        }
-
-        if maxSquareDistance > tolerance {
-            if index - first > 1 {
-                simplifyDouglasPeuckerStep(coordinates, first: first, last: index, tolerance: tolerance, simplified: &simplified)
-            }
-            simplified.append(coordinates[index])
-            if last - index > 1 {
-                simplifyDouglasPeuckerStep(coordinates, first: index, last: last, tolerance: tolerance, simplified: &simplified)
-            }
-        }
-    }
-
-    private func simplifyDouglasPeucker(_ coordinates: [LocationCoordinate2D], tolerance: Double) -> [LocationCoordinate2D] {
-        if coordinates.count <= 2 {
-            return coordinates
-        }
-
-        let lastPoint = coordinates.count - 1
-        var result = [coordinates[0]]
-        simplifyDouglasPeuckerStep(coordinates, first: 0, last: lastPoint, tolerance: tolerance, simplified: &result)
-        result.append(coordinates[lastPoint])
-        return result
-    }
-
     /// Returns a copy of the LineString with the Ramer–Douglas–Peucker algorithm applied to it.
     ///
     /// tolerance:  Controls the level of simplification by specifying the maximum allowed distance between the original line point
@@ -331,14 +241,6 @@ extension LineString {
     ///
     /// Ported from https://github.com/Turfjs/turf/blob/master/packages/turf-simplify/lib/simplify.js
     public mutating func simplified(tolerance: Double = 1.0, highestQuality: Bool = false) {
-        guard coordinates.count > 2 else { return }
-
-        let squareTolerance = tolerance * tolerance
-
-        if !highestQuality {
-            simplified(radialTolerance: squareTolerance)
-        }
-
-        coordinates = simplifyDouglasPeucker(coordinates, tolerance: squareTolerance)
+        coordinates = Simplifier.simplify(coordinates, tolerance: tolerance, highestQuality: highestQuality)
     }
 }
